refactor: removed print statements and changed the  from u32 to u8

+++ src/sort/bubble_sort.rs
@@ -1,6 +1,6 @@
 #[allow(dead_code)]
 #[allow(unused_variables)]
-pub fn bubble_sort(arr: &mut [u32]) {
+pub fn bubble_sort(arr: &mut [u8]) {
     loop {
         let mut swapped: bool = false;
